msgblock: Move message manipulation code from serialqueue.c to new msgblock.c
